Move examples into derive reference
