+++ examples/derive_ref/4_flatten_hand_args.rs
